+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Sankey uses the right total value to calculate percentages when there are more than one sample.
+
 ## [v0.5.3]
 ### Fixed
 - Pin pendatic to <2.0.0 for conda.